// src/api/models/ProduitResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProduitResponse = {
    idProduit?: string;
    nomProduit?: string;
    typeProduit?: string;
    prixVente?: number;
    cout?: number;
    categorie?: string;
    reference?: string;
    codeBarre?: string;
    photo?: string;
    active?: boolean;
    createdAt?: string;
    updatedAt?: string;
};